cdk/scripts_ray/requirements.txt
langchain
ray[data]
sentence-transformers
tqdm
torch
transformers
opensearch-py
psycopg2-binary 
pgvector
pinecone-client